Fix C++ format specifier warnings for int64_t values
- Change %ld to %lld for int64_t and long values in format strings
- Add explicit casts to long long for portable formatting across architectures
- Fixes warnings in MyService.cpp (basicTypes, complexType, returnComplexType)
- Fixes warnings in ComplexType.h (toString method)

This ensures correct formatting on both 32-bit and 64-bit Android architectures
where int64_t may be defined as either 'long' or 'long long'.

+++ Common/src/main/cpp/includes/ComplexType.h
@@ -76,8 +76,8 @@ public:
     {
         char strBuf[1024];
         snprintf(strBuf, 1024,
-                "ComplexType{int=%d, long=%ld, bool=%d, float=%f, double=%lf, string='%s'}",
-                 i_Int, l_Long, b_Boolean, f_Float, d_Double, s_String.c_str());
+                "ComplexType{int=%d, long=%lld, bool=%d, float=%f, double=%lf, string='%s'}",
+                 i_Int, (long long)l_Long, b_Boolean, f_Float, d_Double, s_String.c_str());
 
         pOutputString->append(strBuf);
     }
